+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 *Note: Numbers like (\#123) point to closed Pull Requests on the fractal-web repository.*
 
-# Unreleased
+# 0.10.2
 
 * Added search functionality on jobs table filters (\#424).